+++ web/satellite/src/components/header/projectSelection/ProjectSelectionDropdown.vue
@@ -26,7 +26,7 @@ import { BUCKET_ACTIONS } from '@/store/modules/buckets';
 import { PROJECTS_ACTIONS } from '@/store/modules/projects';
 import { PROJECT_USAGE_ACTIONS } from '@/store/modules/usage';
 import { Project } from '@/types/projects';
-import { setSelectedProjectId } from '@/utils/consoleLocalStorage';
+import { setSelectedProjectId } from '@/utils/localData';
 import {
     API_KEYS_ACTIONS,
     APP_STATE_ACTIONS,
