+++ web/satellite/src/utils/localData.ts
